updatedAt fixes to show title or content updation
